Add simulated Object Storage Service tests based on the previous live tests
