+++ src/OpenStackNetTests.Unit/OpenStackNetTests.Unit.net45.csproj
@@ -43,6 +43,9 @@
     <AssemblyOriginatorKeyFile>..\testing\SDKTestingKey.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="ICSharpCode.SharpZipLib">
+      <HintPath>..\packages\SharpZipLib.0.86.0\lib\20\ICSharpCode.SharpZipLib.dll</HintPath>
+    </Reference>
     <Reference Include="Newtonsoft.Json, Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\packages\Newtonsoft.Json.6.0.4\lib\net45\Newtonsoft.Json.dll</HintPath>
@@ -73,6 +76,9 @@
     <Compile Include="..\OpenStackNetTests.Live\IdentityV2Tests.Shared.cs">
       <Link>IdentityV2Tests.Shared.cs</Link>
     </Compile>
+    <Compile Include="..\OpenStackNetTests.Live\ObjectStorageTests.Shared.cs">
+      <Link>ObjectStorageTests.Shared.cs</Link>
+    </Compile>
     <Compile Include="..\OpenStackNetTests.Live\TestCredentials.cs">
       <Link>TestCredentials.cs</Link>
     </Compile>
@@ -84,6 +90,7 @@
     </Compile>
     <Compile Include="BaseIdentityTests.cs" />
     <Compile Include="IdentityV2Tests.cs" />
+    <Compile Include="ObjectStorageTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -128,6 +135,12 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>IdentityServiceResources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="..\testing\integration\Providers\Rackspace\CloudFiles\DarkKnightRises.jpg">
+      <Link>DarkKnightRises.jpg</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
