+++ src/OpenStackNetTests.Unit/packages.config
@@ -2,4 +2,5 @@
 <packages>
   <package id="Newtonsoft.Json" version="6.0.4" targetFramework="net45" />
   <package id="Rackspace.Net.UriTemplate" version="1.0.0-beta002" targetFramework="net45" />
+  <package id="SharpZipLib" version="0.86.0" targetFramework="net45" />
 </packages>
